requirements.txt
# Core dependencies
pyspark==3.2.0
delta-spark==2.0.0
torch==1.9.0
numpy==1.21.0
pandas==1.3.0

# Data processing
pyarrow==6.0.1
fastparquet==0.8.1

# Machine Learning
scikit-learn==0.24.2
tensorboard==2.6.0

# Utilities
tqdm==4.62.2
python-dotenv==0.19.0
pyyaml==5.4.1

# Testing
pytest==6.2.5
pytest-spark==0.6.0

# Jupyter
jupyter==1.0.0
notebook==6.4.0 